Move exceptions to own folder.
